requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.10.3=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asteroid-filterbanks=0.4.0=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
audioread=3.0.0=pypi_0
av=10.0.0=pypi_0
backports-cached-property=1.0.2=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cmaes=0.9.1=pypi_0
cmake=3.26.3=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
colorlog=6.7.0=pypi_0
commonmark=0.9.1=pypi_0
contourpy=1.0.7=pypi_0
ctranslate2=3.12.0=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cudnn=8.2.1=cuda11.3_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
docopt=0.6.2=pypi_0
einops=0.3.2=pypi_0
faster-whisper=0.4.1=pypi_0
ffmpeg=4.2.2=h20bf706_0
ffmpeg-python=0.2.0=py_0
filelock=3.12.0=pypi_0
flatbuffers=23.3.3=pypi_0
fonttools=4.39.3=pypi_0
freetype=2.10.4=h0708190_1
frozenlist=1.3.3=pypi_0
fsspec=2023.4.0=pypi_0
future=0.18.3=pyhd8ed1ab_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.17.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
greenlet=2.0.2=pypi_0
grpcio=1.54.0=pypi_0
hmmlearn=0.2.8=pypi_0
huggingface-hub=0.13.4=pypi_0
humanfriendly=10.0=pypi_0
hyperpyyaml=1.2.0=pypi_0
idna=3.4=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pypi_0
julius=0.2.7=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h5eee18b_0
librosa=0.9.2=pypi_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
lit=16.0.1=pypi_0
llvmlite=0.39.1=pypi_0
mako=1.2.4=pypi_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
more-itertools=9.1.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.6=he412f7d_0
networkx=2.8.8=pypi_0
numba=0.56.4=pypi_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
onnxruntime=1.14.1=pypi_0
openai-whisper=20230314=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
optuna=3.1.1=pypi_0
packaging=23.1=pypi_0
pandas=2.0.0=pypi_0
pillow=9.5.0=pypi_0
pip=23.0.1=py310h06a4308_0
platformdirs=3.2.0=pypi_0
pooch=1.7.0=pypi_0
primepy=1.3=pypi_0
protobuf=3.20.1=pypi_0
pyannote-audio=2.1.1=pypi_0
pyannote-core=4.5=pypi_0
pyannote-database=4.1.3=pypi_0
pyannote-metrics=3.2.1=pypi_0
pyannote-pipeline=2.3=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.2=pypi_0
pyee=9.0.4=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
python=3.10.11=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytorch-lightning=1.6.5=pypi_0
pytorch-metric-learning=1.7.3=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2023.3.23=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.4.2=pypi_0
rich=12.6.0=pypi_0
rsa=4.9=pypi_0
ruamel-yaml=0.17.21=pypi_0
ruamel-yaml-clib=0.2.7=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
semver=2.13.0=pypi_0
sentencepiece=0.1.98=pypi_0
setuptools=66.0.0=py310h06a4308_0
shellingham=1.5.0.post1=pypi_0
simplejson=3.19.1=pypi_0
singledispatchmethod=1.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sortedcontainers=2.4.0=pypi_0
sounddevice=0.4.6=pypi_0
soundfile=0.10.3.post1=pypi_0
speechbrain=0.5.14=pypi_0
sqlalchemy=2.0.9=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.12.2=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tiktoken=0.3.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
torch=1.13.1=pypi_0
torch-audiomentations=0.11.0=pypi_0
torch-pitch-shift=1.2.4=pypi_0
torchaudio=0.13.1=pypi_0
torchmetrics=0.11.4=pypi_0
tqdm=4.65.0=pypi_0
triton=2.0.0=pypi_0
typer=0.7.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.15=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py310h06a4308_0
whisper-ctranslate2=0.1.9=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.10=h5eee18b_1
yarl=1.8.2=pypi_0
zlib=1.2.13=h5eee18b_0
